First commit for GESS
This is my code for the game of GESS (contraction of Go and Chess, more info in the readme). I have been working on this a few hours a week since early october.

It's my first commit to GitHub ; I'd been working without version control for this project. Here is the source code as a zip, and the game as a jar.

For those unfamiliar with Java, the game can be run with the command "java -jar GESS_0.4.2.jar" in a Unix command line.